Update - component: CardList
- Moved types from component to dedicated folder
